fix(clickhouse): scan channel_info/task JSON as string and ship mini routing
ClickHouse returns String columns as Go strings; the old []byte-only
channel_info Scan failed InitChannelCache and left the distributor with
no channels. Harden JSON scanners, keep Seedance Mini routing/pricing,
and expose roboneo-style task content routes.

+++ controller/channel.go
@@ -719,7 +719,7 @@ type ChannelTag struct {
 	Tag            string  `json:"tag"`
 	NewTag         *string `json:"new_tag"`
 	Priority       *int64  `json:"priority"`
-	Weight         *uint   `json:"weight"`
+	Weight         *uint64 `json:"weight"`
 	ModelMapping   *string `json:"model_mapping"`
 	Models         *string `json:"models"`
 	Groups         *string `json:"groups"`
